initial commit with some of the bootloader working
execution currently loops at 0x20002250 which seems to be waiting for AES to become ready. maybe it was given something to process? I'm not sure what to do about AES yet... I think for now the UID and GID will just be zeros and I'll feed it images signed with that if necessary. Or just make AES no-op since it probably doesn't matter. SHA1 should probably be implemented, though.

Some TODOs:
 - SHA1 peripheral doesn't do anything
 - SPI doesn't do anything and probably should