aes and sha1 might be implemented, image header verification is failing
currently, verify_img_header @ 0x200005dc is failing because I have the wrong file
for the efi bootloader (I need a whole EFI image with the contents decrypted!)
verify_img_header is returning 0 which reutrns to boot_spi @ 0x20003084. if it were
to return 1, then it would verify and decrypt the image and proceed to jump to it.