fixed some things, added paravirtualization interfaces
I might be using the word "paravirtualization" wrong - I'm using it here to mean
something that speeds up virtualization by not simulating a full hardware peripheral.
what I've found is that reading the entire EFI from simulated NOR takes forever. so I
am stubbing out that whole routine so that it copies data straight from simulated NOR's
memory into the virtual s5l8702's memory. this is instant and makes bootup very fast.
the current issue now is diassembling THUMB instructions in the output (and obviously
other things as well) but that's a good problem because WE'RE IN THE EFI BOOTLOADER!!!